Refs #16672 Move project loading code into class.
